+++ src/error/wrap.rs
@@ -118,3 +118,12 @@ mod ext {
         }
     }
 }
+
+pub(crate) mod private {
+    use super::*;
+
+    pub trait Sealed {}
+
+    impl<T, E> Sealed for Result<T, E> where E: ext::StdError {}
+    impl<T> Sealed for Option<T> {}
+}